requirements.txt
aiolimiter==1.2.1
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
Django==5.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
nest-asyncio==1.6.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-telegram-bot==22.1
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.13.2
tzlocal==5.3.1
whitenoise==6.9.0
